# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cruiz API",
    "description": "API для управления продуктами и пользователями",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "schemes": [
    "http"
  ],
  "paths": {
    "/products": {
      "get": {
        "summary": "Получить список продуктов",
        "responses": {
          "200": {
            "description": "Список продуктов успешно получен"
          }
        }
      },
      "post": {
        "summary": "Создать новый продукт",
        "responses": {
          "201": {
            "description": "Продукт успешно создан"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Получить информацию о продукте по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID продукта"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о продукте успешно получена"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Получить список пользователей",
        "responses": {
          "200": {
            "description": "Список пользователей успешно получен"
          }
        }
      },
      "post": {
        "summary": "Добавить нового пользователя",
        "responses": {
          "201": {
            "description": "Пользователь успешно добавлен"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Получить информацию о пользователе по ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID пользователя"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о пользователе успешно получена"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "Уникальный идентификатор продукта"
        },
        "title": {
          "type": "string",
          "description": "Название продукта"
        },
        "count": {
          "type": "integer",
          "description": "Стоимость продукта"
        },
        "description": {
          "type": "string",
          "description": "Описание продукта"
        },
        "image": {
          "type": "string",
          "description": "URL изображения продукта"
        },
        "cruiseDate": {
          "type": "string",
          "description": "даты начало и окончание круиза"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "Уникальный идентификатор пользователя"
        },
        "surName": {
          "type": "string",
          "description": "Фамилия пользователя"
        },
        "firstName": {
          "type": "string",
          "description": "Имя пользователя"
        },
        "middleName": {
          "type": "string",
          "description": "Отчество пользователя"
        },
        "phoneNumber": {
          "type": "string",
          "format": "string",
          "description": "Номер телефона пользователя"
        },
        "numberOfPeople": {
          "type": "integer",
          "description": "Количество человек"
        },
        "cabinNumber": {
          "type": "integer",
          "description": "Номер каюты"
        },
        "productId": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "description": "Идентификатор продукта"
            },
            "title": {
              "type": "string",
              "description": "Название круиза"
            },
            "cruiseDate": {
              "type": "string",
              "description": "даты начало и окончание круиза"
            },
            "count": {
              "type": "integer",
              "description": "Стоимость круиза"
            }
          }
        }
      }
    }
  }
}